refactor: update uses of `TestSigners`
